+++ daemon_rmq_2_sysmon.py
@@ -23,8 +23,6 @@
 """
 
 # Libraries and Global Variables
-from __future__ import print_function
-from __future__ import absolute_import
 
 # Standard
 import sys
@@ -40,9 +38,9 @@ try:
     from . import version
 
 except (ValueError, ImportError) as err:
-    import lib.gen_libs as gen_libs
-    import lib.gen_class as gen_class
-    import rmq_2_sysmon
+    import lib.gen_libs as gen_libs                     # pylint:disable=R0402
+    import lib.gen_class as gen_class                   # pylint:disable=R0402
+    import rmq_2_sysmon                                 # pylint:disable=R0402
     import version
 
 __version__ = version.__version__
@@ -95,7 +93,7 @@ def is_active(pidfile, proc_name):
 
     status = False
 
-    with open(pidfile, "r") as pfile:
+    with open(pidfile, mode="r", encoding="UTF-8") as pfile:
         pid = int(pfile.read().strip())
 
     if pid:
@@ -142,8 +140,8 @@ def main():
 
             if os.path.isfile(pid_file) and is_active(pid_file, proc_name):
 
-                print("Warning:  Pidfile %s exists and process is running."
-                      % (pid_file))
+                print(f"Warning:  Pidfile {pid_file} exists and process is"
+                      f" possibly running.")
 
             elif os.path.isfile(pid_file):
                 os.remove(pid_file)
@@ -165,9 +163,8 @@ def main():
         sys.exit(0)
 
     else:
-        print("Usage: %s -a start|stop|restart -c module -d directory/config \
-{rmq_2_sysmon options}"
-              % sys.argv[0])
+        print(f"Usage: {sys.argv[0]} -a start|stop|restart -c module"
+              f" -d directory/config <rmq_2_sysmon_options_here>")
         sys.exit(2)