+++ test/integration/rmq_2_sysmon/run_program.py
@@ -21,15 +21,15 @@ import mock
 
 # Local
 sys.path.append(os.getcwd())
-import rmq_2_sysmon
-import rmq_cleanup
-import lib.gen_libs as gen_libs
-import version
+import rmq_cleanup                              # pylint:disable=E0401,C0413
+import rmq_2_sysmon                             # pylint:disable=E0401,C0413
+import lib.gen_libs as gen_libs             # pylint:disable=E0401,C0413,R0402
+import version                                  # pylint:disable=E0401,C0413
 
 __version__ = version.__version__
 
 
-class ArgParser(object):
+class ArgParser():
 
     """Class:  ArgParser
 
@@ -53,7 +53,7 @@ class ArgParser(object):
         """
 
         self.cmdline = None
-        self.args = dict()
+        self.args = {}
 
     def get_args_keys(self):
 
@@ -136,7 +136,9 @@ class UnitTest(unittest.TestCase):
         mock_base.return_value = self.test_path
         rmq_2_sysmon.run_program(self.args, self.func_names)
 
-        self.assertTrue(self.connect_true in open(self.cfg.log_file).read())
+        self.assertIn(
+            self.connect_true, open(                    # pylint:disable=R1732
+                self.cfg.log_file, encoding="UTF-8").read())
 
     def tearDown(self):